STYLE:のdefaultがStyles.Doubleになってた

+++ TJADotNet/TJADotNet.cs
@@ -375,7 +375,7 @@ namespace TJADotNet
                                     course.Info.Style = Styles.Double;
                                     break;
                                 default:
-                                    course.Info.Style = Styles.Double;
+                                    course.Info.Style = Styles.Single;
                                     break;
                             }
                         }
